Fixed a lot of errors in sensors.h
The last push had a lot of errors with the multipliers as well as a a lot of missing semicolons. This also included getting data that did not match the sample data we were given. Went in and fixed all of the errors as well as the multipliers, removed case 14 and replaced it with default since there is no pump sensor. Finally, moved 2 variables and made them static so that the could be accessed where they needed to be accessed.